+++ R/gg-miss-fct.R
@@ -26,13 +26,16 @@ gg_miss_fct <- function(x, fct){
   fct <- rlang::enquo(fct)
 
   data <- x %>%
+    # protect against error where grouping by missing value leads to
+    # warning message from dplyr about explicit
+    dplyr::mutate_at(vars(!!fct), .funs = coerce_fct_na_explicit) %>%
     dplyr::group_by(!!fct) %>%
     miss_var_summary()
 
   ggobject <-
     ggplot(data,
-           aes_string(quo_name(fct),
-                      "variable",
+           aes_string(x = quo_name(fct),
+                      y = "variable",
                       fill = "pct_miss")) +
     geom_tile() +
     viridis::scale_fill_viridis(name = "% Miss") +